backend/requirements.txt
# First install build tools
setuptools>=65.0.0
wheel>=0.37.0

# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OpenAI and AI dependencies
openai>=1.3.0
tiktoken>=0.5.0

# Vector database (use latest compatible version)
chromadb>=0.4.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.0
httpx>=0.25.0
loguru>=0.7.0

# Data processing (use compatible versions)
numpy>=1.21.0
pandas>=1.3.0

# Sentence transformers (install separately if needed)
# sentence-transformers>=2.2.0

# LangChain (optional - can be installed separately)
# langchain>=0.0.340
# langchain-openai>=0.0.2